Change /transactions to reflect newest spec

+++ src/main/java/io/swagger/model/Transaction.java
@@ -30,7 +30,7 @@ public class Transaction {
   @NotNull
   @Id
   @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "t_seq")
-  private Integer id = null;
+  private Long id = null;
 
   @Schema(required = true, description = "")
   @NotNull
